+++ html/components/ft-accountCreate.html
@@ -1,0 +1,72 @@
+<div class="col-md-12">
+  <ul class="nav flex-column">
+    <li class="nav-item"><a class="nav-link text-uppercase" [ngClass]="{ 'active': tabs == 1 }" (click)="changeTabs(1)" href="javascript:void(0);">My Contracts</a></li>  
+    <li class="nav-item"><a class="nav-link text-uppercase" [ngClass]="{ 'active': tabs == 2 }" (click)="changeTabs(2)" href="javascript:void(0);">Make a Token</a></li>
+    <li class="nav-item"><a class="nav-link text-uppercase" [ngClass]="{ 'active': tabs == 3 }" (click)="changeTabs(3)" href="javascript:void(0);">Deploy Code - Advanced</a></li>
+  </ul>
+  <div class="tab-white tab-content text-left">
+    <div class="tab-pane fade show roboto padtabs" [ngClass]="{ 'active': tabs == 1 }">
+        <strong>My Contracts</strong>
+        <br><Br>
+        <div *ngIf="myContracts">
+          <span *ngFor="let contract of myContracts">{{contract.ContractNameVer}}<br>
+          <small>{{contract.PublishedAddress}}<br><Br></small></span>
+        </div>
+    </div>
+    <div class="tab-pane fade show roboto padtabs" [ngClass]="{ 'active': tabs == 2 }">
+      <label><strong>Token Name</strong></label>
+      <input id="tokenName" [(ngModel)]="tokenName" (keyup)="removeNonLetters();" placeholder="TheBestExampleCoin">
+      <br><br>
+      <div *ngIf="!publishedAddress">
+        <a class="btn btn-lg btn-ft" (click)="this.name();compile();" href="javascript:void(0);">
+          <span *ngIf="compiledCode">Re</span>Compile Code
+        </a> - 
+        <a *ngIf="compiledCode" class="btn btn-lg btn-ft" (click)="publish();" href="javascript:void(0);">
+          Publish "{{compiledName}}"
+          <small *ngIf="gasEst"><br>Gas Est: {{gasEst | number}}</small>
+        </a>
+      </div>
+      <div *ngIf="publishedAddress">
+        <label><strong>Contract Name: </strong></label><br>
+        {{compiledName}}<br><br>
+        <label><strong>Contract Published:</strong></label><br>
+        {{publishedAddress}} <br><br>
+      </div>
+      <label *ngIf="compiledCode"><br><strong>Compiled Code:</strong><br></label>
+      <pre *ngIf="compiledCode">{{compiledCode}}</pre>
+      <label *ngIf="compiledCode"><br><strong>ABI:</strong><br></label>
+      <pre *ngIf="compiledCode" [innerHTML]="JSON.stringify(compiledABI)"></pre>
+    </div>
+    <div class="tab-pane fade show roboto padtabs" [ngClass]="{ 'active': tabs == 3 }">
+      <div>
+        <label><strong>Paste Solidity Smart Contract</strong></label><br>
+        <textarea style="width:100%; height:400px;" id="solidityCode" spellcheck="false" [(ngModel)]="solidityCode">
+        </textarea>
+        <br>
+        <label *ngIf="compileError"><br><strong>Error Compiling:</strong><br> - Use this tool to publish solidity contracts. 
+          For development, we recomend using <a href="https://remix.ethereum.org/" rel="nofollow" target="_blank">remix</a>.<br></label>
+          <pre *ngIf="compileError">{{compileError}}</pre>
+        <br>
+        <label *ngIf="compiledCode"><br><strong>Compiled Code:</strong><br></label>
+          <pre *ngIf="compiledCode">{{compiledCode}}</pre>
+          <label *ngIf="compiledCode"><br><strong>ABI:</strong><br></label>
+          <pre *ngIf="compiledCode" [innerHTML]="JSON.stringify(compiledABI)"></pre>
+      
+        <div *ngIf="!publishedAddress">
+          <a class="btn btn-lg btn-ft" (click)="compile();" href="javascript:void(0);">
+            <span *ngIf="compiledCode">Re</span>Compile Code
+          </a> - <a *ngIf="compiledCode" class="btn btn-lg btn-ft" (click)="publish();" href="javascript:void(0);">
+              Publish "{{compiledName}}"
+              <small *ngIf="gasEst"><br>Gas Est: {{gasEst | number}}</small>
+          </a><br>
+        </div>
+        <div *ngIf="publishedAddress">
+          <label><strong>Contract Name: </strong></label><br>
+          {{compiledName}}<br><br>
+          <label><strong>Contract Published:</strong></label><br>
+          {{publishedAddress}} <br><br>
+        </div>
+      </div>        
+    </div>
+  </div>
+</div>
