+++ html/components/ft-accountExit.html
@@ -1,0 +1,8 @@
+<h4>
+  <strong>E</strong>xit
+</h4>
+<div><br>
+  <a class="btn btn-lg btn-ft" (click)="signOut();" href="javascript:void(0);">
+      Sign Out
+  </a>
+</div>
