+++ html/modal.html
@@ -180,7 +180,7 @@
                 </h3>
                 <p>
                     <span>
-                        Yes. We publish all blocks from the FinTechToken network on S3 every minute so anyone can verify transactions.
+                        Yes. We <a href="https://s3.amazonaws.com/fintechtoken.chain/" rel="nofollow" target="_blank">publish</a> all blocks from the FinTechToken network on S3 every 10 minutes so anyone can verify transactions.
                         <br><Br>
                         Our code is Token Sourced and available for anyone to read.
                         <br><Br>
